added navigation links and navbar
