+++ package.json
@@ -11,7 +11,10 @@
     "nanoid": "^3.1.18",
     "popper.js": "^1.16.1",
     "react": "^17.0.1",
+    "react-bootstrap-validation": "^0.1.11",
     "react-dom": "^17.0.1",
+    "react-router-bootstrap": "^0.25.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
     "recharts": "^1.8.5",
     "web-vitals": "^0.2.4"
