+++ src/components/Home.js
@@ -1,0 +1,12 @@
+import React from 'react';
+ 
+const home = () => {
+    return (
+       <div>
+          <h1>Home</h1>
+           <p>Home page body content</p>
+       </div>
+    );
+}
+ 
+export default home;
