+++ src/components/design.css
@@ -1,0 +1,23 @@
+.design {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    background-color: white;
+    padding: 20px;
+    box-shadow: 0 4px 8px #00000030;
+   
+    }
+
+a {
+    display: block;
+    color: black;
+}
+a:active{
+    color: grey;
+}
+a:hover{
+    color: grey;
+}
+a:link{
+    text-decoration: none;
+}
